test-requirements.txt
coverage>=3.7.1
docker ~= 2.1.0
flake8
pytest < 3.3
pytest-bdd
pytest-cov
PyYAML
dictdiffer
